starting with the john files
gonna be honest i completely forgot how to make custom folders so this is just gonna sit pretty in the middle of the space until i remember how. my only experience with github was a shitty attempt at a roblox repaint. yeah i know it was stupid and good luck trying to find it. the only thing i remember was that making new folders is a nightmare